requirements.txt
flask>=3.0.0
jinja2>=3.1.0
python-docx>=1.1.2
